backend/requirements.txt
django==4.0.6
Pillow==9.2.0
django-environ==0.9.0
PyMySQL==1.0.2
mysql-connector-python==8.0.23
setuptools==58.1.0
mysqlclient==2.1.1
djangorestframework==3.13.1
#django-filter==22.1
#django-rest-swagger==2.2.0
bcrypt==3.2.2
pygments
